+++ src/librustc_mir/transform/deaggregator.rs
@@ -10,8 +10,9 @@
 
 use rustc::ty::TyCtxt;
 use rustc::mir::*;
-use rustc::mir::transform::{MirPass, MirSource};
+use rustc::mir::transform::MirSource;
 use rustc_data_structures::indexed_vec::Idx;
+use transform::MirPass;
 
 pub struct Deaggregator;
