+++ src/librustc_mir/transform/simplify_branches.rs
@@ -12,8 +12,9 @@
 
 use rustc::ty::{self, TyCtxt};
 use rustc::middle::const_val::ConstVal;
-use rustc::mir::transform::{MirPass, MirSource};
+use rustc::mir::transform::MirSource;
 use rustc::mir::*;
+use transform::MirPass;
 
 use std::borrow::Cow;
